feat: simplify dashboard, add settings page, profile avatar dropdown
Dashboard now only shows: Stats, Log Activity, Activity Heatmap,
Embeddable Badge for GitHub, Profile Views, Your Projects, and Inbox.

Moved Edit Profile, Email Preferences, and Referral to new /settings page.

Replaced navbar logout button with circular profile avatar dropdown
showing: My Profile, Settings, Referral, and Log out options.
